docker_example/tesseract/requirements.txt
opencv-python
numpy
pillow
pytesseract
flask
flask-cors
gunicorn
